+++ KlayGE/Samples/src/Parallax/Parallax.cpp
@@ -107,7 +107,7 @@ namespace
 						if (rf.RenderEngineInstance().DeviceCaps().a2bgr10_vertex_support)
 						{
 							init_data.data = &normals[0];
-							init_data.row_pitch = normals.size() * sizeof(normals[0]);
+							init_data.row_pitch = static_cast<uint32_t>(normals.size() * sizeof(normals[0]));
 							init_data.slice_pitch = init_data.row_pitch;
 							GraphicsBufferPtr normal_vb = rf.MakeVertexBuffer(BU_Static, EAH_GPU_Read, &init_data);
 							rl_->SetVertexStream(i, normal_vb);
@@ -116,7 +116,7 @@ namespace
 						else
 						{
 							init_data.data = &normals_float3[0];
-							init_data.row_pitch = normals_float3.size() * sizeof(normals_float3[0]);
+							init_data.row_pitch = static_cast<uint32_t>(normals_float3.size() * sizeof(normals_float3[0]));
 							init_data.slice_pitch = init_data.row_pitch;
 							GraphicsBufferPtr normal_vb = rf.MakeVertexBuffer(BU_Static, EAH_GPU_Read, &init_data);
 							rl_->SetVertexStream(i, normal_vb);
@@ -150,7 +150,7 @@ namespace
 						if (rf.RenderEngineInstance().DeviceCaps().a2bgr10_vertex_support)
 						{
 							init_data.data = &tangents[0];
-							init_data.row_pitch = tangents.size() * sizeof(tangents[0]);
+							init_data.row_pitch = static_cast<uint32_t>(tangents.size() * sizeof(tangents[0]));
 							init_data.slice_pitch = init_data.row_pitch;
 							GraphicsBufferPtr tangent_vb = rf.MakeVertexBuffer(BU_Static, EAH_GPU_Read, &init_data);
 							rl_->SetVertexStream(i, tangent_vb);
@@ -159,7 +159,7 @@ namespace
 						else
 						{
 							init_data.data = &tangents_float3[0];
-							init_data.row_pitch = tangents_float3.size() * sizeof(tangents_float3[0]);
+							init_data.row_pitch = static_cast<uint32_t>(tangents_float3.size() * sizeof(tangents_float3[0]));
 							init_data.slice_pitch = init_data.row_pitch;
 							GraphicsBufferPtr tangent_vb = rf.MakeVertexBuffer(BU_Static, EAH_GPU_Read, &init_data);
 							rl_->SetVertexStream(i, tangent_vb);
